Isolate generic API tool execution from Onshape dispatch
